refactor: Resolve pylint issues
