+++ backend/capellacollab/core/authentication/provider/azure/keystore.py
@@ -1,6 +1,8 @@
 # SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
 # SPDX-License-Identifier: Apache-2.0
 
+# pylint: skip-file
+
 from __future__ import annotations
 
 import logging
