+++ backend/capellacollab/core/database/migration.py
@@ -47,7 +47,9 @@ def migrate_db(engine, database_url: str):
             "script_location", str(root_dir / "alembic")
         )
         alembic_cfg.set_main_option("sqlalchemy.url", database_url)
-        alembic_cfg.attributes["configure_logger"] = False
+        alembic_cfg.attributes[  # pylint:disable=unsupported-assignment-operation
+            "configure_logger"
+        ] = False
 
         with engine.connect() as conn:
             context = MigrationContext.configure(conn)
